Reorganized Code for Readability
Worked on reorganizing the code into partial classes so that it's a little easier to read and understand.  Made edits to the acquire_images thread to account for the user having a longer loop sequence than the desired number of frames.

Program crashed at one point.  Need to work on recreating this to see why.