Edit Interface Updates
Updated the Edit options to activate on double-clicking a CurrentSettings object.  Editing now includes a "Save" button color change to alert the user to the change in functionality.  Additional minor variable change in the FileHandling.cs file to make code a little more readable and easy to edit in the future.

Version updated to 0.2.2.  This version works on argos-dev-2 with minimal issue.  Writes fits files per frame.