First Exposure Timing Shuttering Edit
Capture algorithm changed to take advantage of the Acquire and Run functions.  Interface update to reflect new functionality.  Many more changes to come.  This is the first commit of the latest capture process:  exposure times will be changed to match the time it takes to rotate the filter wheel between frames, and then the bad frame will be removed post-acquisition.