Acquisition Filter File IO and Minor Updates
SaveCurrentFilterSettings() method has been updated to take an optional filename parameter.  This was added to aid in the next edit:

On Run/Acquire press (not the transition from Run to Acquire or reverse), the current filter settings are written out to a .dat file with the same name as the .spe file, plus "_FilterSettings".

The information on TriggerSlewAdjust is now stored in the save file for CurrentSettings list, and read in appropriately.