First Frame Error Fix and Partial Logging
Implemented logging for manual control.  More updates to come for automated control, and some formatting updates in the log file itself.

Increased the thread sleep time in the SetNextExpTime() method in the Automated event handler methods section of ControlPanel.xaml.cs to be 800ms for regular frames, 1000ms for the first frame.  This corrected the first frame errors we were seeing.